Refactor: Move Toast component to App.tsx and remove ReloadProvider
This commit refactors the location of the Toast component and removes the ReloadProvider.

The Toast component was moved from `src/UI/screens/rotinas/index.tsx` to `App.tsx` to ensure it's globally accessible.
The ReloadProvider was removed from `src/UI/screens/rotinas/index.tsx` as it's no longer needed with the Toast component being globally available.
A navigation button to Rotinas was added in EixosScreen for testing purposes.